v0.3.6 Update
